docker: Get agent plugin running again (WIP, check changes still to do)
As the agent plugin now only returns one set of CPU data counters,
(note the use of one_shot=True) usage calculations can no longer be done
in the agent and will have to be handled in the check plugin. The calculated
value has been removed and replaced by the usage counters for the container and
total respectively.

In the docker images section, we removed all information calculated from
container information, as this can also be done in the check plugin.